Structure End Child Poverty Act parameters as a reform Fixes #3476
