+++ policyengine_us/parameters/gov/irs/credits/refundable.yaml
@@ -24,11 +24,9 @@ values:
     - recovery_rebate_credit
     - refundable_payroll_tax_credit
     - premium_tax_credit
-    - ecpa_filer_credit
-    - ecpa_adult_dependent_credit
 metadata:
   unit: program
-  name: refundable_tax_credits
+  period: year
   label: Refundable tax credits
   reference:
     - title: 26 U.S. Code Subpart C - Refundable Credits
