+++ policyengine_us/reforms/reforms.py
@@ -11,6 +11,11 @@ from .cbo.payroll import (
     create_increase_taxable_earnings_for_social_security_reform,
 )
 from .congress.wyden_smith import create_ctc_expansion_reform
+
+from .congress.tlaib.end_child_poverty_act import create_ecpa_adult_dependent_credit_reform
+from .congress.tlaib.end_child_poverty_act import create_ecpa_filer_credit_reform
+
+
 from policyengine_core.reforms import Reform
 import warnings
 
@@ -34,6 +39,10 @@ def create_structural_reforms_from_parameters(parameters, period):
     )
     ctc_expansion = create_ctc_expansion_reform(parameters, period)
 
+    ecpa_adult_dependent_credit = create_ecpa_adult_dependent_credit_reform(parameters, period)
+
+    ecpa_filer_credit = create_ecpa_filer_credit_reform(parameters, period)
+
     reforms = [
         afa_reform,
         winship_reform,
@@ -42,7 +51,10 @@ def create_structural_reforms_from_parameters(parameters, period):
         remove_head_of_household,
         increase_taxable_earnings_for_social_security_reform,
         ctc_expansion,
+        ecpa_adult_dependent_credit,
+        ecpa_filer_credit,
     ]
+
     reforms = tuple(filter(lambda x: x is not None, reforms))
 
     class combined_reform(Reform):
